:recycle: Refactor: 비즈니스 로직 이동
<body>
Util 이 아인 TagGenerator 서비스층으로 이동

- 관련 : #343
