+++ application/external-api-v2/src/main/java/com/namo/spring/application/external/api/guest/service/GuestManageService.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Service;
 
 import com.namo.spring.application.external.api.guest.dto.GuestParticipantRequest;
 import com.namo.spring.application.external.api.schedule.service.ParticipantMaker;
-import com.namo.spring.application.external.global.utils.TagGenerator;
+import com.namo.spring.application.external.api.user.service.TagGenerator;
 import com.namo.spring.core.common.code.status.ErrorStatus;
 import com.namo.spring.db.mysql.domains.category.entity.Palette;
 import com.namo.spring.db.mysql.domains.category.exception.PaletteException;
