+++ application/external-api-v2/src/main/java/com/namo/spring/application/external/api/user/service/TagGenerator.java
@@ -1,14 +1,10 @@
-package com.namo.spring.application.external.global.utils;
+package com.namo.spring.application.external.api.user.service;
 
 import java.util.HashSet;
 import java.util.Random;
 import java.util.Set;
 
 import org.springframework.stereotype.Component;
-
-import com.namo.spring.application.external.api.user.service.MemberManageService;
-import com.namo.spring.core.common.code.status.ErrorStatus;
-import com.namo.spring.db.mysql.domains.user.exception.MemberException;
 
 import lombok.RequiredArgsConstructor;
