+++ application/external-api-v2/src/test/java/com/namo/spring/application/external/api/user/service/TagGeneratorTest.java
@@ -12,8 +12,6 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
-import com.namo.spring.application.external.global.utils.TagGenerator;
-
 public class TagGeneratorTest {
 
     @Mock
